aipr_py_agent/requirements.txt
google-generativeai
python-dotenv
rich
duckduckgo-search
